% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciftest.R
\name{get_weightit_mparts}
\alias{get_weightit_mparts}
\title{Extract Mparts from a weightit object (robustly)}
\usage{
get_weightit_mparts(weightit)
}
\description{
Extract Mparts from a weightit object (robustly)
}
