% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciftest-phase2.R
\name{predict_censoring_km}
\alias{predict_censoring_km}
\title{Evaluate a censoring Kaplan-Meier distribution}
\usage{
predict_censoring_km(
  object,
  time,
  strata = rep.int("pooled", length(time)),
  side = c("left", "right"),
  use = c("used", "raw")
)
}
\description{
Evaluate a censoring Kaplan-Meier distribution
}
\keyword{internal}
