% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciftest-phase2.R
\name{predict_working_aj}
\alias{predict_working_aj}
\title{Evaluate a working Aalen-Johansen distribution}
\usage{
predict_working_aj(
  object,
  time,
  exposure,
  strata = rep.int("pooled", length(time)),
  side = c("left", "right"),
  use = c("used", "raw")
)
}
\description{
Evaluate a working Aalen-Johansen distribution
}
\keyword{internal}
