% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifplot.R
\name{call_ggsurvfit}
\alias{call_ggsurvfit}
\title{Plot survival or cumulative incidence curves with ggsurvfit}
\usage{
call_ggsurvfit(
  survfit_object,
  out_readSurv = NULL,
  conf.type = NULL,
  addConfidenceInterval = TRUE,
  addRiskTable = TRUE,
  addEstimateTable = FALSE,
  addCensorMark = TRUE,
  shape.censor.mark = 3,
  size.censor.mark = 2,
  addCompetingRiskMark = TRUE,
  competing.risk.time = list(),
  shape.competing.risk.mark = 16,
  size.competing.risk.mark = 2,
  addIntercurrentEventMark = TRUE,
  intercurrent.event.time = list(),
  shape.intercurrent.event.mark = 1,
  size.intercurrent.event.mark = 2,
  addQuantileLine = FALSE,
  quantile = 0.5,
  type.y = NULL,
  label.x = "Time",
  label.y = NULL,
  label.strata = NULL,
  limits.x = NULL,
  limits.y = NULL,
  breaks.x = NULL,
  breaks.y = NULL,
  use_coord_cartesian = FALSE,
  style = "CLASSIC",
  font.family = "sans",
  font.size = 14,
  legend.position = "top"
)
}
\arguments{
\item{survfit_object}{A \code{survfit} object.}

\item{out_readSurv}{(optional) List returned by your \code{readSurv()} to auto-set x limits.}

\item{conf.type}{Character transformation for CI on the probability scale (default \code{"arcsine-square root"}).}

\item{addConfidenceInterval}{Logical add \code{add_confidence_interval()} to plot. It calls geom_ribbon() (default \code{TRUE}).}

\item{addRiskTable}{Logical add \code{add_risktable(risktable_stats="n.risk")} to plot (default \code{TRUE}).}

\item{addEstimateTable}{Logical add \code{add_risktable(risktable_stats="estimate (conf.low, conf.high)")} to plot (default \code{FALSE}).}

\item{addCensorMark}{Logical add \code{add_censor_mark()} to plot. It calls geom_point() (default \code{TRUE}).}

\item{shape.censor.mark}{Integer point shape for censor marks (default \code{3}).}

\item{size.censor.mark}{Numeric point size for censor marks (default \code{2}).}

\item{addCompetingRiskMark}{Logical add time marks to describe event2 specified by Event(), usually the competing events. It calls geom_point() (default \code{TRUE}).}

\item{shape.competing.risk.mark}{Integer point shape for competing-risk marks (default \code{16}).}

\item{size.competing.risk.mark}{Numeric point size for competing-risk marks (default \code{2}).}

\item{addIntercurrentEventMark}{Logical overlay user-specified time marks per strata calls geom_point() (default \code{TRUE}).}

\item{intercurrent.event.time}{Named list of numeric vectors (names must be mapped to strata labels).}

\item{shape.intercurrent.event.mark}{Integer point shape for intercurrent-event marks (default \code{1}).}

\item{size.intercurrent.event.mark}{Numeric point size for intercurrent-event marks (default \code{2}).}

\item{addQuantileLine}{Logical add \code{add_quantile()} to plot. It calls geom_segment() (default \code{TRUE}).}

\item{quantile}{Numeric specify quantile for \code{add_quantile()} (default \code{0.5}).}

\item{type.y}{\code{NULL} (survival) or \code{"risk"} (display \code{1 - survival} i.e. CIF).}

\item{label.x}{Character x-axis labels (default \code{"Time"}).}

\item{label.y}{Character y-axis labels (default internally set to \code{"Survival"} or \code{"Cumulative incidence"}).}

\item{label.strata}{Character vector of labels for strata.}

\item{limits.x}{Numeric length-2 vectors for axis limits. If NULL it is internally set to \code{c(0,max(out_readSurv$t))}.}

\item{limits.y}{Numeric length-2 vectors for axis limits. If NULL it is internally set to \code{c(0,1)}.}

\item{breaks.x}{Numeric vectors for axis breaks (default \code{NULL}).}

\item{breaks.y}{Numeric vectors for axis breaks (default \code{NULL}).}

\item{use_coord_cartesian}{Logical specify use of coord_cartesian() (default \code{FALSE}).}

\item{style}{Character plot theme controls (default \code{"CLASSIC"}).}

\item{font.family}{Character plot theme controls (default \code{"sans"}).}

\item{font.size}{Integer plot theme controls (default \code{14}).}

\item{legend.position}{Character specify position of legend: \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"}, or \code{"none"} (default \code{"top"}).}
}
\value{
A \code{ggplot} object.
Plot survival or cumulative incidence curves with ggsurvfit
...
...
}
\description{
Plot survival or cumulative incidence curves with ggsurvfit
}
