% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifcurve.R
\name{call_ggsurvfit}
\alias{call_ggsurvfit}
\title{Plot survival or cumulative incidence curves with ggsurvfit}
\usage{
call_ggsurvfit(
  survfit_object,
  out_readSurv = NULL,
  ggsurvfit.type = NULL,
  conf.type = "arcsine-square root",
  addConfidenceInterval = TRUE,
  addRiskTable = TRUE,
  addCensorMark = TRUE,
  shape.censor.mark = 3,
  size.censor.mark = 2,
  addIntercurrentEventMark = TRUE,
  intercurrent.event.time = list(),
  shape.intercurrent.event.mark = 16,
  size.intercurrent.event.mark = 2,
  label.x = "Time",
  label.y = "Survival probability",
  lims.x = NULL,
  lims.y = c(0, 1),
  font.family = "sans",
  font.size = 14,
  legend.position = "top"
)
}
\arguments{
\item{survfit_object}{A \code{survfit} object.}

\item{out_readSurv}{(optional) List returned by your \code{readSurv()} to auto-set x limits.}

\item{ggsurvfit.type}{Character; NULL (survival) or "risk" for cumulative incidence display.}

\item{conf.type}{Character; same as used when constructing CI (e.g., "none", "n", "arcsine-square root").}

\item{addConfidenceInterval}{Logical; add CI via \code{add_confidence_interval()}.}

\item{addRiskTable}{Logical; add risk table via \code{add_risktable(risktable_stats="n.risk")}.}

\item{addCensorMark}{Logical; add censor marks via \code{add_censor_mark()}.}

\item{shape.censor.mark}{Integer; shape for censor marks. Defaults to 3.}

\item{size.censor.mark}{Numeric; size for censor marks. Defaults to 2.}

\item{addIntercurrentEventMark}{Logical; add user-specified marks at strata-specific times.}

\item{intercurrent.event.time}{Named list of numeric vectors.}

\item{shape.intercurrent.event.mark}{Integer; fixed shape for intercurrent marks across strata. Defaults 16.}

\item{size.intercurrent.event.mark}{Numeric; size for intercurrent marks. Defaults 2.}

\item{label.x, label.y}{Axis labels. If \code{label.y = "Survival probability"} (default) and
\code{ggsurvfit.type == "risk"}, it is automatically replaced with \code{"1 - survival probability"}.}

\item{lims.x}{Numeric length-2; x limits. If NULL and \code{out_readSurv} given, uses \code{c(0,max(out_readSurv$t))}.}

\item{lims.y}{Numeric length-2; y limits.}

\item{font.family, font.size}{Theme controls.}

\item{legend.position}{"top","right","bottom","left" or "none".}
}
\value{
A \code{ggplot} object.
}
\description{
Plot survival or cumulative incidence curves with ggsurvfit
}
